refactor(M6 PR2): extract git-state application into controller/git_apply.rs
Move set_changed / apply_git_state / refresh_git_state / drop_pending_status and
handle_focus_gained into a submodule. The re-root orchestration that drives them
stays in mod.rs. Cross-dispatch methods pub(super). Behavior-preserving.
mod.rs ~1860.
